ported packages, gpp to x64 windows. Still there are a number of warnings about convertion from prolog_int to int and signed/unsigned mismatch between int and size_t on win32
git-svn-id: svn+ssh://svn.code.sf.net/p/xsb/src/trunk@7176 756869e0-9e47-4bc5-b0a8-847305dc5e1c

+++ XSB/packages/curl/cc/load_page.c
@@ -147,7 +147,7 @@ load_page (char *source, curl_opt options, curl_ret *ret_vals)
 void *
 encode (char *url, char **dir, char **file, char **suffix)
 {
-  int dir_len = 0;
+  size_t dir_len = 0;
   char *dir_enc = NULL, *file_enc = NULL, *suff_enc = NULL, *ptr;
 
   ptr = strrchr (url, '/');
